requirements.txt
alembic
aniso8601
bcrypt
cffi
click
Flask
Flask-Bcrypt
Flask-Migrate
flask-restplus
Flask-Script
Flask-SQLAlchemy
Flask-Testing
itsdangerous
Jinja2
jsonschema
Mako
MarkupSafe
pycparser
PyJWT
python-dateutil
python-editor
pytz
six
SQLAlchemy
Werkzeug
pandas
requests
mysql-connector-python
python-dotenv
DrissionPage
google-cloud-vision
webdriver-manager

